+++ docs/.vitepress/config.js
@@ -67,6 +67,7 @@ export default {
                     { text: "Calico NetworkPolicy", link: "/showcases/calico-network-policies.html" },
                     { text: "Cluster API (CAPI)", link: "/showcases/cluster-api.html" },
                     { text: "Crossplane Platform Engineering", link: "/showcases/crossplane.html" },
+                    { text: "CloudNativePG", link: "/showcases/cloudnative-pg.html" },
                     { text: "Headlamp (Dashboard)", link: "/showcases/headlamp.html" }
                 ]
             },